refactor: extract key event handler from mod.rs into input_handler.rs
mod.rs was 1278 lines with a monolithic key event handler chain.
Extracted all key input handling into input_handler.rs with clear
function decomposition:

- handle_key() as top-level dispatcher
- Per-mode handlers: help_overlay, queue_execution, completion_summary,
  continuous_pause, session_summary, log_viewer, confirm_kill,
  session_switcher
- Global shortcuts: F-keys, Ctrl-X, help toggle
- Overview/default mode keys with grid navigation

mod.rs is now 608 lines (52% reduction), focused on the event loop,
terminal setup, command dispatch, and session lifecycle.
