+++ src/tui/mod.rs
@@ -7,6 +7,7 @@ pub mod detail;
 pub mod fullscreen;
 pub mod help;
 pub mod icons;
+mod input_handler;
 pub mod log_viewer;
 pub mod markdown;
 pub mod marquee;
@@ -30,16 +31,13 @@ mod snapshot_tests;
 use crate::github::client::{GhCliClient, GitHubClient};
 use crate::tui::activity_log::LogLevel;
 use app::App;
-use background_tasks::{spawn_issue_fetch, spawn_upgrade_download, spawn_version_check};
+use background_tasks::{spawn_issue_fetch, spawn_version_check};
 use crossterm::{
-    event::{
-        self, DisableMouseCapture, EnableMouseCapture, Event, KeyCode, KeyModifiers, MouseEventKind,
-    },
+    event::{self, DisableMouseCapture, EnableMouseCapture, Event, KeyCode, MouseEventKind},
     execute,
     terminal::{EnterAlternateScreen, LeaveAlternateScreen, disable_raw_mode, enable_raw_mode},
 };
 use ratatui::{Terminal, backend::CrosstermBackend};
-use screen_dispatch::{dispatch_to_active_screen, handle_screen_action};
 use std::io;
 use std::time::Duration;
 use summary::print_summary;
@@ -100,700 +98,31 @@ async fn event_loop(
         if event::poll(Duration::from_millis(50))? {
             match event::read()? {
                 Event::Key(key) => {
-                    match &app.upgrade_state {
-                        crate::updater::UpgradeState::Available(info) => {
-                            if key.code == KeyCode::Char('u') {
-                                let info_clone = info.clone();
-                                let tx = app.data_tx.clone();
-                                app.upgrade_state = crate::updater::UpgradeState::Downloading {
-                                    version: info_clone.version.clone(),
-                                };
-                                spawn_upgrade_download(tx, info_clone);
-                                continue;
-                            }
-                            if key.code == KeyCode::Esc {
-                                app.upgrade_state = crate::updater::UpgradeState::Hidden;
-                                continue;
-                            }
-                        }
-                        crate::updater::UpgradeState::ReadyToRestart { .. } => {
-                            if key.code == KeyCode::Char('y') {
-                                disable_raw_mode().ok();
-                                execute!(
-                                    terminal.backend_mut(),
-                                    LeaveAlternateScreen,
-                                    DisableMouseCapture
-                                )
-                                .ok();
-                                terminal.show_cursor().ok();
-                                if let Err(e) = crate::updater::installer::restart_with_same_args()
-                                {
-                                    enable_raw_mode().ok();
-                                    execute!(
-                                        io::stdout(),
-                                        EnterAlternateScreen,
-                                        EnableMouseCapture
-                                    )
-                                    .ok();
-                                    app.upgrade_state = crate::updater::UpgradeState::Failed(
-                                        format!("Restart failed: {}", e),
-                                    );
-                                }
-                                continue;
-                            }
-                            if key.code == KeyCode::Char('n') || key.code == KeyCode::Esc {
-                                app.upgrade_state = crate::updater::UpgradeState::Hidden;
-                                app.activity_log.push_simple(
-                                    "UPDATE".into(),
-                                    "Upgrade installed. Restart manually to use new version."
-                                        .into(),
-                                    crate::tui::activity_log::LogLevel::Info,
-                                );
-                                continue;
-                            }
-                        }
-                        crate::updater::UpgradeState::Failed(_) => {
-                            if key.code == KeyCode::Esc || key.code == KeyCode::Enter {
-                                app.upgrade_state = crate::updater::UpgradeState::Hidden;
-                                continue;
-                            }
-                        }
-                        _ => {}
-                    }
-
-                    if let Some(ref mut help) = app.help_state {
-                        if help.search_active {
-                            match key.code {
-                                KeyCode::Esc => help.clear_search(),
-                                KeyCode::Enter => {
-                                    help.search_active = false;
-                                }
-                                KeyCode::Backspace => help.pop_char(),
-                                KeyCode::Char(c) => help.push_char(c),
-                                _ => {}
-                            }
-                        } else {
-                            match key.code {
-                                KeyCode::Char('?') | KeyCode::Esc | KeyCode::F(1) => {
-                                    app.help_state = None;
-                                }
-                                KeyCode::Char('j') | KeyCode::Down => help.scroll_down(),
-                                KeyCode::Char('k') | KeyCode::Up => help.scroll_up(),
-                                KeyCode::PageDown => help.page_down(),
-                                KeyCode::PageUp => help.page_up(),
-                                KeyCode::Char('/') => help.toggle_search(),
-                                _ => {}
-                            }
-                        }
-                        continue;
-                    }
-
-                    if app.tui_mode == app::TuiMode::QueueExecution {
-                        use crate::work::executor::{ExecutorPhase, FailureAction};
-                        match (key.code, key.modifiers) {
-                            (KeyCode::Esc, _) => {
-                                app.tui_mode = app::TuiMode::Overview;
-                            }
-                            (KeyCode::Char('q'), _)
-                            | (KeyCode::Char('c'), KeyModifiers::CONTROL) => {
-                                app.running = false;
-                                return Ok(());
-                            }
-                            (KeyCode::Char('r'), _) => {
-                                if let Some(ref mut exec) = app.queue_executor
-                                    && matches!(
-                                        exec.phase(),
-                                        ExecutorPhase::AwaitingDecision { .. }
-                                    )
-                                {
-                                    exec.apply_decision(FailureAction::Retry);
-                                    app.advance_queue_and_launch();
-                                }
-                            }
-                            (KeyCode::Char('s'), _) => {
-                                if let Some(ref mut exec) = app.queue_executor
-                                    && matches!(
-                                        exec.phase(),
-                                        ExecutorPhase::AwaitingDecision { .. }
-                                    )
-                                {
-                                    exec.apply_decision(FailureAction::Skip);
-                                    if exec.is_finished() {
-                                        app.completion_summary =
-                                            Some(app.build_completion_summary());
-                                        app.tui_mode = app::TuiMode::CompletionSummary;
-                                    } else {
-                                        app.advance_queue_and_launch();
-                                    }
-                                }
-                            }
-                            (KeyCode::Char('a'), _) => {
-                                if let Some(ref mut exec) = app.queue_executor
-                                    && matches!(
-                                        exec.phase(),
-                                        ExecutorPhase::AwaitingDecision { .. }
-                                    )
-                                {
-                                    exec.apply_decision(FailureAction::Abort);
-                                    app.completion_summary = Some(app.build_completion_summary());
-                                    app.tui_mode = app::TuiMode::CompletionSummary;
-                                }
-                            }
-                            _ => {}
-                        }
-                        continue;
-                    }
-
-                    if app.tui_mode == app::TuiMode::CompletionSummary {
-                        match (key.code, key.modifiers) {
-                            (KeyCode::Enter, _) | (KeyCode::Esc, _) => {
-                                app.transition_to_dashboard();
-                            }
-                            (KeyCode::Char('q'), _)
-                            | (KeyCode::Char('c'), KeyModifiers::CONTROL) => {
-                                app.running = false;
-                                return Ok(());
-                            }
-                            (KeyCode::Char('i'), _) => {
-                                app.completion_summary = None;
-                                app.completion_summary_dismissed = true;
-                                let mut screen = screens::IssueBrowserScreen::new(vec![]);
-                                screen.loading = true;
-                                app.issue_browser_screen = Some(screen);
-                                app.pending_commands.push(app::TuiCommand::FetchIssues);
-                                app.tui_mode = app::TuiMode::IssueBrowser;
-                            }
-                            (KeyCode::Char('r'), _) => {
-                                app.prompt_input_screen = Some(
-                                    app::helpers::create_prompt_input_screen(&app.prompt_history),
-                                );
-                                app.tui_mode = app::TuiMode::PromptInput;
-                            }
-                            (KeyCode::Char('l'), _) => {
-                                if let Some(ref summary) = app.completion_summary {
-                                    if let Some(first) = summary.sessions.first() {
-                                        let sid = first.session_id;
-                                        app.log_viewer_scroll = 0;
-                                        app.completion_summary = None;
-                                        app.tui_mode = app::TuiMode::LogViewer(sid);
-                                    } else {
-                                        app.tui_mode = app::TuiMode::Overview;
-                                    }
-                                } else {
-                                    app.tui_mode = app::TuiMode::Overview;
-                                }
-                            }
-                            (KeyCode::Char('f'), _) => {
-                                let has_suggestions = app
-                                    .completion_summary
-                                    .as_ref()
-                                    .map(|s| s.has_conflict_suggestions())
-                                    .unwrap_or(false);
-                                if has_suggestions {
-                                    let config = app.completion_summary.as_ref().and_then(|s| {
-                                        s.suggestions.get(s.selected_suggestion).map(|sg| {
-                                            screens::ConflictFixConfig {
-                                                pr_number: sg.pr_number,
-                                                issue_number: sg.issue_number,
-                                                branch: sg.branch.clone(),
-                                                conflicting_files: sg.conflicting_files.clone(),
-                                            }
-                                        })
-                                    });
-                                    if let Some(config) = config {
-                                        app.spawn_conflict_fix_session(&config);
-                                        app.completion_summary = None;
-                                        app.tui_mode = app::TuiMode::Overview;
-                                    }
-                                } else {
-                                    let needs_review: Vec<_> = app
-                                        .completion_summary
-                                        .as_ref()
-                                        .into_iter()
-                                        .flat_map(|s| &s.sessions)
-                                        .filter(|s| {
-                                            s.status
-                                                == crate::session::types::SessionStatus::NeedsReview
-                                        })
-                                        .cloned()
-                                        .collect();
-                                    if !needs_review.is_empty() {
-                                        for sl in &needs_review {
-                                            app.spawn_gate_fix_session(sl);
-                                        }
-                                        app.completion_summary = None;
-                                        app.tui_mode = app::TuiMode::Overview;
-                                    }
-                                }
-                            }
-                            (KeyCode::Up, _) | (KeyCode::Char('k'), _) => {
-                                if let Some(ref mut summary) = app.completion_summary
-                                    && !summary.suggestions.is_empty()
-                                {
-                                    summary.selected_suggestion =
-                                        summary.selected_suggestion.saturating_sub(1);
-                                }
-                                app.panel_view.scroll_up();
-                            }
-                            (KeyCode::Down, _) | (KeyCode::Char('j'), _) => {
-                                if let Some(ref mut summary) = app.completion_summary
-                                    && !summary.suggestions.is_empty()
-                                {
-                                    let max = summary.suggestions.len().saturating_sub(1);
-                                    if summary.selected_suggestion < max {
-                                        summary.selected_suggestion += 1;
-                                    }
-                                }
-                                app.panel_view.scroll_down();
-                            }
-                            _ => {}
-                        }
-                        if !matches!(
-                            key.code,
-                            KeyCode::Up | KeyCode::Down | KeyCode::Char('k' | 'j')
-                        ) {
-                            app.completion_summary = None;
-                        }
-                        continue;
-                    }
-
-                    if app.tui_mode == app::TuiMode::ContinuousPause {
-                        match (key.code, key.modifiers) {
-                            (KeyCode::Char('s'), _) => {
-                                if let Some(ref mut cont) = app.continuous_mode {
-                                    let skipped = cont.current_failure().map(|f| f.issue_number);
-                                    cont.on_skip();
-                                    if let Some(num) = skipped {
-                                        app.activity_log.push_simple(
-                                            "CONTINUOUS".into(),
-                                            format!("Skipped #{}, advancing...", num),
-                                            LogLevel::Warn,
-                                        );
-                                    }
-                                }
-                                app.tui_mode = app::TuiMode::Overview;
-                            }
-                            (KeyCode::Char('r'), _) => {
-                                if let Some(ref mut cont) = app.continuous_mode
-                                    && let Some(issue_number) = cont.on_retry()
-                                {
-                                    if let Some(ref mut assigner) = app.work_assigner {
-                                        assigner.mark_pending_undo_cascade(issue_number);
-                                    }
-                                    app.activity_log.push_simple(
-                                        "CONTINUOUS".into(),
-                                        format!("Retrying #{}...", issue_number),
-                                        LogLevel::Info,
-                                    );
-                                }
-                                app.tui_mode = app::TuiMode::Overview;
-                            }
-                            (KeyCode::Char('q'), _)
-                            | (KeyCode::Char('c'), KeyModifiers::CONTROL) => {
-                                app.continuous_mode = None;
-                                app.running = false;
-                                return Ok(());
-                            }
-                            _ => {}
-                        }
-                        continue;
-                    }
-
-                    // Only trigger `?` help when not in a text input mode;
-                    // F1 always works. (#280 fix: `?` was stealing input in PromptInput)
-                    let is_text_input_mode = matches!(
-                        app.tui_mode,
-                        app::TuiMode::PromptInput | app::TuiMode::SessionSwitcher
-                    );
-                    if key.code == KeyCode::F(1)
-                        || (key.code == KeyCode::Char('?') && !is_text_input_mode)
+                    if let crate::updater::UpgradeState::ReadyToRestart { .. } = &app.upgrade_state
+                        && key.code == KeyCode::Char('y')
                     {
-                        app.help_state = Some(crate::tui::help::HelpOverlayState::new());
-                        continue;
-                    }
-
-                    // F-key aliases for DOS-style function key bar (#218)
-                    if let KeyCode::F(n) = key.code {
-                        match n {
-                            2 => {
-                                let summary = app.build_completion_summary();
-                                app.completion_summary = Some(summary);
-                                app.session_summary_state =
-                                    Some(crate::tui::app::types::SessionSummaryState::default());
-                                app.tui_mode = app::TuiMode::SessionSummary;
-                                continue;
-                            }
-                            3 => {
-                                let selected = app.panel_view.selected_index();
-                                if let Some(id) = app.pool.session_id_at_index(selected) {
-                                    app.tui_mode = app::TuiMode::Fullscreen(id);
-                                }
-                                continue;
-                            }
-                            4 => {
-                                app.tui_mode = app::TuiMode::CostDashboard;
-                                continue;
-                            }
-                            5 => {
-                                app.tui_mode = app::TuiMode::TokenDashboard;
-                                continue;
-                            }
-                            6 => {
-                                app.tui_mode = match app.tui_mode {
-                                    app::TuiMode::Overview => app::TuiMode::DependencyGraph,
-                                    app::TuiMode::DependencyGraph => app::TuiMode::CostDashboard,
-                                    app::TuiMode::CostDashboard => app::TuiMode::TokenDashboard,
-                                    app::TuiMode::TokenDashboard => app::TuiMode::Overview,
-                                    _ => app::TuiMode::Overview,
-                                };
-                                continue;
-                            }
-                            #[cfg(unix)]
-                            9 => {
-                                app.pause_all();
-                                continue;
-                            }
-                            10 => {
-                                let selected = app.panel_view.selected_index();
-                                if let Some(id) = app.pool.session_id_at_index(selected)
-                                    && let Some(session) = app.pool.get_session(id)
-                                    && !session.status.is_terminal()
-                                {
-                                    app.tui_mode = app::TuiMode::ConfirmKill(id);
-                                }
-                                continue;
-                            }
-                            _ => {}
-                        }
-                    }
-
-                    // Ctrl-X to exit (#218)
-                    if key.code == KeyCode::Char('x')
-                        && key.modifiers.contains(KeyModifiers::CONTROL)
-                    {
-                        app.running = false;
-                        return Ok(());
-                    }
-
-                    let event = Event::Key(key);
-                    let screen_handled =
-                        if let Some(action) = dispatch_to_active_screen(app, &event) {
-                            handle_screen_action(app, action);
-                            true
-                        } else {
-                            false
-                        };
-
-                    if screen_handled {
-                        if !app.running {
-                            return Ok(());
+                        disable_raw_mode().ok();
+                        execute!(
+                            terminal.backend_mut(),
+                            LeaveAlternateScreen,
+                            DisableMouseCapture
+                        )
+                        .ok();
+                        terminal.show_cursor().ok();
+                        if let Err(e) = crate::updater::installer::restart_with_same_args() {
+                            enable_raw_mode().ok();
+                            execute!(io::stdout(), EnterAlternateScreen, EnableMouseCapture).ok();
+                            app.upgrade_state = crate::updater::UpgradeState::Failed(format!(
+                                "Restart failed: {}",
+                                e
+                            ));
                         }
                         continue;
                     }
 
-                    // Handle SessionSummary input
-                    if app.tui_mode == app::TuiMode::SessionSummary {
-                        match (key.code, key.modifiers) {
-                            (KeyCode::Esc, _) => {
-                                app.session_summary_state = None;
-                                app.tui_mode = app::TuiMode::Overview;
-                            }
-                            (KeyCode::Char('q'), _)
-                            | (KeyCode::Char('c'), KeyModifiers::CONTROL) => {
-                                app.running = false;
-                                return Ok(());
-                            }
-                            (KeyCode::Up, _) | (KeyCode::Char('k'), _) => {
-                                if let Some(state) = app.session_summary_state.as_mut() {
-                                    if state.selected_index > 0 {
-                                        state.selected_index -= 1;
-                                    }
-                                    state.scroll_up();
-                                }
-                            }
-                            (KeyCode::Down, _) | (KeyCode::Char('j'), _) => {
-                                let max_idx = app
-                                    .completion_summary
-                                    .as_ref()
-                                    .map(|s| s.sessions.len().saturating_sub(1))
-                                    .unwrap_or(0);
-                                if let Some(state) = app.session_summary_state.as_mut() {
-                                    if state.selected_index < max_idx {
-                                        state.selected_index += 1;
-                                    }
-                                    state.scroll_down();
-                                }
-                            }
-                            (KeyCode::Enter, _) => {
-                                let session_id = app.completion_summary.as_ref().and_then(|s| {
-                                    let idx = app
-                                        .session_summary_state
-                                        .as_ref()
-                                        .map(|st| st.selected_index)
-                                        .unwrap_or(0);
-                                    s.sessions.get(idx).map(|sl| sl.session_id)
-                                });
-                                if let (Some(id), Some(state)) =
-                                    (session_id, app.session_summary_state.as_mut())
-                                {
-                                    state.toggle_expand(id);
-                                }
-                            }
-                            _ => {}
-                        }
-                        continue;
-                    }
-
-                    // Handle LogViewer input
-                    if let app::TuiMode::LogViewer(id) = app.tui_mode {
-                        match (key.code, key.modifiers) {
-                            (KeyCode::Esc, _) => {
-                                app.tui_mode = app::TuiMode::Detail(id);
-                            }
-                            (KeyCode::Char('q'), _)
-                            | (KeyCode::Char('c'), KeyModifiers::CONTROL) => {
-                                app.running = false;
-                                return Ok(());
-                            }
-                            (KeyCode::Up, _) | (KeyCode::Char('k'), _) => {
-                                app.log_viewer_scroll = app.log_viewer_scroll.saturating_sub(1);
-                            }
-                            (KeyCode::Down, _) | (KeyCode::Char('j'), _) => {
-                                app.log_viewer_scroll = app.log_viewer_scroll.saturating_add(1);
-                            }
-                            (KeyCode::Char('G'), _) => {
-                                app.log_viewer_scroll = u16::MAX;
-                            }
-                            (KeyCode::Char('g'), _) => {
-                                app.log_viewer_scroll = 0;
-                            }
-                            _ => {}
-                        }
-                        continue;
-                    }
-
-                    // Handle ConfirmKill input
-                    if let app::TuiMode::ConfirmKill(session_id) = app.tui_mode {
-                        match key.code {
-                            KeyCode::Char('y') | KeyCode::Enter => {
-                                app.kill_selected_session(session_id).await;
-                                app.tui_mode = app::TuiMode::Overview;
-                            }
-                            KeyCode::Char('n') | KeyCode::Esc => {
-                                app.tui_mode = app::TuiMode::Overview;
-                            }
-                            _ => {}
-                        }
-                        continue;
-                    }
-
-                    // Handle SessionSwitcher input
-                    if app.tui_mode == app::TuiMode::SessionSwitcher {
-                        match key.code {
-                            KeyCode::Esc => {
-                                app.session_switcher = None;
-                                app.tui_mode = app::TuiMode::Overview;
-                            }
-                            KeyCode::Up => {
-                                if let Some(sw) = &mut app.session_switcher {
-                                    sw.move_up();
-                                }
-                            }
-                            KeyCode::Down => {
-                                if let Some(sw) = &mut app.session_switcher {
-                                    let count = {
-                                        let sessions = app.pool.all_sessions();
-                                        let refs: Vec<&crate::session::types::Session> = sessions;
-                                        sw.filtered_sessions(&refs).len()
-                                    };
-                                    sw.move_down(count);
-                                }
-                            }
-                            KeyCode::Enter => {
-                                let selected_id = app.session_switcher.as_ref().and_then(|sw| {
-                                    let sessions = app.pool.all_sessions();
-                                    sw.selected_session(&sessions).map(|s| s.id)
-                                });
-                                if let Some(id) = selected_id {
-                                    app.session_switcher = None;
-                                    app.tui_mode = app::TuiMode::Detail(id);
-                                }
-                            }
-                            _ => {}
-                        }
-                        continue;
-                    }
-
-                    match (key.code, key.modifiers) {
-                        (KeyCode::Char('q'), _) | (KeyCode::Char('c'), KeyModifiers::CONTROL) => {
-                            app.running = false;
-                            return Ok(());
-                        }
-                        #[cfg(unix)]
-                        (KeyCode::Char('p'), _) => {
-                            app.pause_all();
-                        }
-                        #[cfg(unix)]
-                        (KeyCode::Char('r'), _) => {
-                            app.resume_all();
-                        }
-                        (KeyCode::Char('k'), _) => {
-                            let selected = app.panel_view.selected_index();
-                            if let Some(id) = app.pool.session_id_at_index(selected)
-                                && let Some(session) = app.pool.get_session(id)
-                                && !session.status.is_terminal()
-                            {
-                                app.tui_mode = app::TuiMode::ConfirmKill(id);
-                            }
-                        }
-                        (KeyCode::Char('K'), _) => {
-                            app.kill_all().await;
-                        }
-                        (KeyCode::Char('f'), _) => {
-                            let selected = app.panel_view.selected_index();
-                            if let Some(id) = app.pool.session_id_at_index(selected) {
-                                app.tui_mode = app::TuiMode::Fullscreen(id);
-                            }
-                        }
-                        (KeyCode::Char('$'), _) => {
-                            app.tui_mode = app::TuiMode::CostDashboard;
-                        }
-                        (KeyCode::Char('t'), _) => {
-                            app.tui_mode = app::TuiMode::TokenDashboard;
-                        }
-                        (KeyCode::Char('S'), _) => {
-                            let summary = app.build_completion_summary();
-                            app.completion_summary = Some(summary);
-                            app.session_summary_state =
-                                Some(crate::tui::app::types::SessionSummaryState::default());
-                            app.tui_mode = app::TuiMode::SessionSummary;
-                        }
-                        (KeyCode::Tab, _) => {
-                            app.tui_mode = match app.tui_mode {
-                                app::TuiMode::Overview => app::TuiMode::DependencyGraph,
-                                app::TuiMode::DependencyGraph => app::TuiMode::CostDashboard,
-                                app::TuiMode::CostDashboard => app::TuiMode::TokenDashboard,
-                                app::TuiMode::TokenDashboard => app::TuiMode::Overview,
-                                app::TuiMode::Detail(_)
-                                | app::TuiMode::Fullscreen(_)
-                                | app::TuiMode::Dashboard
-                                | app::TuiMode::IssueBrowser
-                                | app::TuiMode::MilestoneView
-                                | app::TuiMode::PromptInput
-                                | app::TuiMode::CompletionSummary
-                                | app::TuiMode::ContinuousPause
-                                | app::TuiMode::QueueConfirmation
-                                | app::TuiMode::QueueExecution
-                                | app::TuiMode::HollowRetry
-                                | app::TuiMode::Sanitize
-                                | app::TuiMode::Settings
-                                | app::TuiMode::SessionSwitcher
-                                | app::TuiMode::AdaptWizard
-                                | app::TuiMode::PrReview
-                                | app::TuiMode::ReleaseNotes
-                                | app::TuiMode::LogViewer(_)
-                                | app::TuiMode::ConfirmKill(_)
-                                | app::TuiMode::SessionSummary => app::TuiMode::Overview,
-                            };
-                        }
-                        (KeyCode::Esc, _) => {
-                            if app.home_screen.is_some() && app.pool.total_count() == 0 {
-                                app.tui_mode = app::TuiMode::Dashboard;
-                            } else {
-                                app.tui_mode = app::TuiMode::Overview;
-                            }
-                        }
-                        (KeyCode::Enter, _) | (KeyCode::Char(' '), _) => {
-                            let selected = app.panel_view.selected_index();
-                            if let Some(id) = app.pool.session_id_at_index(selected)
-                                && let Some(session) = app.pool.get_session(id)
-                            {
-                                if session.status.is_terminal() {
-                                    app.toggle_session_summary(id);
-                                } else {
-                                    app.tui_mode = app::TuiMode::Detail(id);
-                                }
-                            }
-                        }
-                        (KeyCode::Char(c), _) if c.is_ascii_digit() && c != '0' => {
-                            let idx = (c as usize) - ('1' as usize);
-                            if let Some(id) = app.pool.session_id_at_index(idx) {
-                                app.tui_mode = app::TuiMode::Detail(id);
-                            }
-                        }
-                        (KeyCode::Char('w'), _) => {
-                            app.session_switcher =
-                                Some(crate::tui::session_switcher::SessionSwitcher::default());
-                            app.tui_mode = app::TuiMode::SessionSwitcher;
-                        }
-                        (KeyCode::Char('d'), _) => {
-                            let selected = app.panel_view.selected_index();
-                            if let Some(id) = app.pool.session_id_at_index(selected)
-                                && let Some(session) = app.pool.get_session(id)
-                                && session.status.is_terminal()
-                            {
-                                app.dismiss_session(id);
-                            } else {
-                                app.notifications.dismiss_latest();
-                            }
-                        }
-                        (KeyCode::Char('D'), _) => {
-                            app.dismiss_all_completed();
-                        }
-                        (KeyCode::Char('l'), _) => match app.tui_mode {
-                            app::TuiMode::Detail(id) => {
-                                app.log_viewer_scroll = 0;
-                                app.tui_mode = app::TuiMode::LogViewer(id);
-                            }
-                            app::TuiMode::Overview => {
-                                let selected = app.panel_view.selected_index();
-                                if let Some(id) = app.pool.session_id_at_index(selected) {
-                                    app.log_viewer_scroll = 0;
-                                    app.tui_mode = app::TuiMode::LogViewer(id);
-                                }
-                            }
-                            _ => {}
-                        },
-                        (KeyCode::Up, KeyModifiers::SHIFT) => {
-                            app.panel_view.scroll_up();
-                        }
-                        (KeyCode::Down, KeyModifiers::SHIFT) => {
-                            app.panel_view.scroll_down();
-                        }
-                        (KeyCode::Up, _)
-                        | (KeyCode::Down, _)
-                        | (KeyCode::Left, _)
-                        | (KeyCode::Right, _)
-                        | (KeyCode::Char('['), _)
-                        | (KeyCode::Char(']'), _) => {
-                            let term_size = terminal.size().unwrap_or_default();
-                            let layout = crate::tui::panels::GridLayout::calculate(
-                                app.pool.total_count(),
-                                term_size.width,
-                                term_size.height,
-                            );
-                            let total_sessions = app.pool.total_count();
-                            match key.code {
-                                KeyCode::Up => app.panel_view.grid_state.move_up(),
-                                KeyCode::Down => {
-                                    app.panel_view.grid_state.move_down(&layout, total_sessions)
-                                }
-                                KeyCode::Left => app.panel_view.grid_state.move_left(),
-                                KeyCode::Right => app
-                                    .panel_view
-                                    .grid_state
-                                    .move_right(&layout, total_sessions),
-                                KeyCode::Char('[') => app.panel_view.grid_state.prev_page(&layout),
-                                KeyCode::Char(']') => app.panel_view.grid_state.next_page(&layout),
-                                _ => {}
-                            }
-                            app.panel_view.selected =
-                                Some(app.panel_view.grid_state.selected_index(&layout));
-                            if !matches!(key.code, KeyCode::Char('[') | KeyCode::Char(']')) {
-                                app.panel_view.scroll_offset = 0;
-                            }
-                        }
-                        _ => {}
+                    match input_handler::handle_key(app, key).await {
+                        input_handler::KeyAction::Consumed => {}
+                        input_handler::KeyAction::Quit => return Ok(()),
                     }
                 }
                 Event::Mouse(mouse) => match mouse.kind {
@@ -1094,6 +423,7 @@ mod handle_screen_action_tests {
     use super::*;
     use crate::session::worktree::MockWorktreeManager;
     use crate::state::store::StateStore;
+    use crate::tui::screen_dispatch::handle_screen_action;
     use screens::ScreenAction;
 
     fn make_app() -> app::App {
